Add AddStudent class
